refactor: 💡 Extract order setup from controller to service
